Change refresh order
Here is a suggestion revision for the order of refresh operations.  This would allow update hooks to be based on recently enabled modules. If an enabled module needs to make revisions outside of its own concerns it still can do this within an install hook.  This solves the problem of trying to deploy and unenabled module (say a contrib module) with a cross cutting update_N hook in a feature perhaps.